requirements.txt
cython
psutil
pyyaml
websockets
gunicorn
gevent
requests
flask
flask-socketio
python-socketio
socketio-client
celery